Streamlit/experiment-setup/requirements.txt
streamlit
pandas
numpy
requests
uvicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
openai